Call the pager with null device instead of testing its existence
The grotty(1) manual suggests

    PAGER="/usr/bin/less -R"
    export PAGER

if that variable is set like this, cmdliner calls

    Sys.command "type /usr/bin/less -R 1>/dev/null 2>/dev/null"

but type(1p) looks for both "/usr/bin/less" and "-R" in the
filesystem. Using -- after type doesn't change the result.

    /usr/bin/less is /usr/bin/less
    -R not found

Another idea is to call directly the pager and pipe the null device
into it. GNU less, more, and Windows more.com return 0 in that case.
It also allows to pass arguments to the pager from cmdliner.